Move event listeners out of io module
